modified the program so that points operate on doubles instead of ints

+++ PolygonDrawer/Polygon/Line.cs
@@ -36,7 +36,7 @@
         {
             get
             {
-                int dx = P1.X - P2.X, dy = P1.Y - P2.Y;
+                double dx = P1.X - P2.X, dy = P1.Y - P2.Y;
                 return Math.Sqrt(dx * dx + dy * dy);
             }
         }
@@ -117,7 +117,7 @@
                                 P[i].L[i].ChangeState(LineState.None);
                             break;
                         case LineState.Vertical:
-                            int avgX = (P[0].X + P[1].X) / 2;
+                            double avgX = (P[0].X + P[1].X) / 2;
                             P[0].MoveLocation(avgX - P[0].X, 0, P[1]);
                             P[1].MoveLocation(avgX - P[1].X, 0, P[0]);
 
@@ -125,7 +125,7 @@
                                 P[i].L[i].ChangeState(LineState.Vertical);
                             break;
                         case LineState.Horizontal:
-                            int avgY = (P[0].Y + P[1].Y) / 2;
+                            double avgY = (P[0].Y + P[1].Y) / 2;
                             P[0].MoveLocation(0, avgY - P[0].Y, P[1]);
                             P[1].MoveLocation(0, avgY - P[1].Y, P[0]);
